+++ src/main/java/com/telepathicgrunt/blame/mixin/StructureMixin.java
@@ -1,0 +1,30 @@
+package com.telepathicgrunt.blame.mixin;
+
+import com.telepathicgrunt.blame.main.StructureFeatureBlame;
+import net.minecraft.util.SharedSeedRandom;
+import net.minecraft.util.math.ChunkPos;
+import net.minecraft.world.gen.feature.structure.Structure;
+import net.minecraft.world.gen.settings.StructureSeparationSettings;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
+
+/* @author - TelepathicGrunt
+ *
+ * Detect who set their structure spacing to 0
+ *
+ * LGPLv3
+ */
+@Mixin(Structure.class)
+public class StructureMixin {
+
+	@Inject(method = "getChunkPosForStructure(Lnet/minecraft/world/gen/settings/StructureSeparationSettings;JLnet/minecraft/util/SharedSeedRandom;II)Lnet/minecraft/util/math/ChunkPos;",
+			at = @At(value = "HEAD"))
+	private void checkSpacing(StructureSeparationSettings separationSettings, long seed, SharedSeedRandom rand, int x, int z, CallbackInfoReturnable<ChunkPos> cir)
+	{
+		if(separationSettings.func_236668_a_() == 0){
+			StructureFeatureBlame.printStructureSpacingBlame((Structure<?>)(Object)this);
+		}
+	}
+}